Add copia puc al crear OrdenPago

+++ Convertidor/Data/Entities/Adm/ADM_PUC_ORDEN_PAGO.cs
@@ -10,9 +10,9 @@
         public int FINANCIADO_ID { get; set; }
         public int CODIGO_FINANCIADO { get; set; }
         public int CODIGO_SALDO { get; set; }
-        public int MONTO { get; set; }
-        public int MONTO_PAGADO { get; set; }
-        public int MONTO_ANULADO { get; set; }
+        public decimal MONTO { get; set; }
+        public decimal MONTO_PAGADO { get; set; }
+        public decimal MONTO_ANULADO { get; set; }
         public string EXTRA1 { get; set; } = string.Empty;
         public string EXTRA2 { get; set; } = string.Empty;
         public string EXTRA3 { get; set; } = string.Empty;
